# [RFC]: Add support for WebSocketSpec