--- a/src/litserve/specs/__init__.py
+++ b/src/litserve/specs/__init__.py
@@ -1,4 +1,5 @@
 from litserve.specs.openai import OpenAISpec
 from litserve.specs.openai_embedding import OpenAIEmbeddingSpec
+from litserve.specs.websocket import WebSocketSpec
 
-__all__ = ["OpenAISpec", "OpenAIEmbeddingSpec"]
+__all__ = ["OpenAISpec", "OpenAIEmbeddingSpec", "WebSocketSpec"]