--- a/src/litserve/specs/base.py
+++ b/src/litserve/specs/base.py
@@ -23,6 +23,7 @@ class LitSpec:
 
     def __init__(self):
         self._endpoints = []
+        self._ws_endpoints = []
 
         self._server: LitServer = None
 
@@ -39,6 +40,10 @@ class LitSpec:
     def add_endpoint(self, path: str, endpoint: Callable, methods: List[str]):
         """Register an endpoint in the spec."""
         self._endpoints.append((path, endpoint, methods))
+
+    def add_ws_endpoint(self, path: str, endpoint: Callable):
+        """Register a websocket endpoint in the spec."""
+        self._ws_endpoints.append((path, endpoint))
 
     @property
     def endpoints(self):