--- a/src/litserve/specs/websocket.py
+++ b/src/litserve/specs/websocket.py
@@ -15,9 +15,9 @@ import asyncio
 import logging
 import time
 import uuid
-from typing import TYPE_CHECKING, Optional, Union
+from typing import TYPE_CHECKING, Any, Dict, Optional
 
-from fastapi import HTTPException, WebSocket, WebSocketDisconnect, status
+from fastapi import WebSocket, WebSocketDisconnect, status
 
 from litserve.specs.base import LitSpec
 from litserve.utils import LitAPIStatus
@@ -40,46 +40,55 @@ class WebSocketSpec(LitSpec):
 
         print("WebSocket Spec is ready.")
 
-    def decode_request(self, request: Union[str, bytes], context_kwargs: Optional[dict] = None) -> str:
-        """Decode the request from the client."""
-        if isinstance(request, bytes):
-            return request.decode("utf-8")
+    def decode_request(self, request: Dict, context_kwargs: Optional[dict] = None) -> Any:
         return request
 
-    def encode_response(self, output: str, context_kwargs: Optional[dict] = None) -> bytes:
-        """Encode the response to send to the client."""
-        if isinstance(output, str):
-            return output.encode("utf-8")
+    def encode_response(self, output: Any, context_kwargs: Optional[dict] = None) -> Dict[str, Any]:
         return output
 
     async def ws_predict(self, websocket: WebSocket):
+        # TODO: Determine if a dedicated connection manager is needed to effectively maintain active WebSocket connections
         await websocket.accept()
+        response_queue_id = self.response_queue_id
+        logger.debug("Received WebSocket connection: %s", websocket.client)
         try:
             while True:
+                # TODO: Discuss support for additional payload formats beyond JSON.
                 payload = await websocket.receive_json()
-                response_queue_id = self.response_queue_id
 
                 uid = uuid.uuid4()
                 event = asyncio.Event()
                 self._server.response_buffer[uid] = event
                 # Send request to inference worker
                 self._server.request_queue.put_nowait((response_queue_id, uid, time.monotonic(), payload))
-                # Wait for response
+                # Wait for the response
                 await event.wait()
                 response, response_status = self._server.response_buffer.pop(uid)
 
                 # Handle errors
                 if response_status == LitAPIStatus.ERROR:
-                    if isinstance(response, HTTPException):
-                        await websocket.send_json({"error": response.detail, "status_code": response.status_code})
-                    else:
-                        await websocket.send_json({"error": "Internal server error", "status_code": 500})
-                else:
-                    # Send successful response back to client
-                    await websocket.send_json(response)
+                    logger.error("Error in WebSocket communication: %s", response)
+                    raise Exception("Error in WebSocket communication")
+
+                logger.debug(response)
+
+                if not isinstance(response, dict):
+                    raise ValueError(
+                        f"Expected response to be a dictionary, but got type {type(response)}.",
+                        "The response should be a dictionary to ensure proper compatibility with the WebSocketSpec.",
+                        "Please ensure that your response is a dictionary.",
+                    )
+
+                # Send successful response back to client
+                await websocket.send_json(response)
 
         except WebSocketDisconnect:
-            logger.debug("WebSocket cliet disconnected")
+            logger.debug("WebSocket client disconnected")
         except Exception as e:
-            logger.error(f"Error in WebSocket communication: {e}")
+            logger.exception("Error in WebSocket communication", exc_info=e)
+            # TODO: Catch unsupported payload formats and send error message
+            await websocket.send_json({"error": "Internal server error", "details": str(e)})
             await websocket.close(code=status.WS_1011_INTERNAL_ERROR)
+        finally:
+            await websocket.close()
+            logger.debug("WebSocket connection closed")