@aniketmaurya 2025-04-22
Hey @bhimrazy, great work adding WebSocket support! 🙌

That said, I noticed that the current implementation still follows the same old flow:
decode request → run prediction → encode response, all in one go.

Right now, we’re still treating the WebSocket like a regular HTTP call, just over a different pipe. It works, but we’re not unlocking the real benefits yet.