@bhimrazy 2025-04-23
Thanks for the thoughtful feedback, @aniketmaurya ! 🙏 
You're absolutely right — there's definitely more we can do to take full advantage of WebSockets. 
I’ll do some more research and also see if we can tweak the Loops flow. Appreciate the nudge! 🚀